backend/requirements.txt
# Flask 核心
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# 数据库
PyMySQL==1.1.0
SQLAlchemy==2.0.23
cryptography==41.0.7

# 密码加密
bcrypt==4.1.1

# 数据验证
email-validator==2.1.0

# 测试
pytest==7.4.3
pytest-cov==4.1.0
hypothesis==6.92.1

# 数据导出
pandas==2.1.4
openpyxl==3.1.2

# 开发工具
python-dotenv==1.0.0
